+++ apps/taskman/exceptions.py
@@ -11,5 +11,9 @@ class MissingJiraTokenException(TaskmanException):
     """exception for performing user action without providing token"""
 
 
+class TaskWritePermissionsException(TaskmanException):
+    """exception for performing task creation/update without permissions in the Jira project"""
+
+
 class JiraTaskErrorException(TaskmanException):
     """exception for getting an error from Jira"""